feat: add indexer feature flag
