+++ crates/bevy-dojo/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-apibara-core = { git = "https://github.com/apibara/dna", rev = "2c452070372e80e5d1a5302c1b8ae5defef7e9a4" }
-apibara-sdk = { git = "https://github.com/apibara/dna", rev = "2c452070372e80e5d1a5302c1b8ae5defef7e9a4" }
+apibara-core = { git = "https://github.com/apibara/dna", rev = "2c452070372e80e5d1a5302c1b8ae5defef7e9a4", optional = true }
+apibara-sdk = { git = "https://github.com/apibara/dna", rev = "2c452070372e80e5d1a5302c1b8ae5defef7e9a4", optional = true }
 bevy = { version = "0.10", default-features = false }
 bevy-tokio-tasks = "0.10"
 chrono = "0.4"
@@ -13,3 +13,12 @@ tokio = { version = "1.27", features = ["full", "sync"] }
 tokio-stream = "0.1"
 tonic = "0.8"
 prost = "0.11"
+
+[features]
+default = ["indexer"]
+indexer = ["apibara-core", "apibara-sdk"]
+
+[[example]]
+name = "simple"
+path = "examples/simple.rs"
+required-features = ["indexer"]
